+++ CHANGELOG.md
@@ -6,6 +6,13 @@ Changes:
 
 \-
 
+## v2023-05-01-1
+
+Changes:
+
+- Transfer changes authentication
+- Only transfer changes if has changes
+
 ## v2023-04-26-1
 
 Changes: